6.10 Change some network functions parameters from int to short to save bandwidth + Fixed a RPC drop item issue

+++ Assets/Scripts/Item/Consumables/HealthPotion.cs
@@ -18,7 +18,7 @@ public class HealthPotion : Consumables
         this.itemType = ItemType.Consumable;
     }
 
-    public HealthPotion(int amount)
+    public HealthPotion(short amount)
     {
         healAmount = 10;
         this.itemName = "HealthPotion";
